Removed deprecated component Library.js
