+++ jukeos/src/App.js
@@ -4,7 +4,6 @@ import { AnimatePresence } from 'framer-motion';
 import NavigationBar from './components/NavigationBar';
 import Home from './components/Home';
 import Library from './components/Library';
-import LibraryTesting from './components/LibraryTesting';
 import Settings from './components/Settings';
 import Harmony from './components/Harmony';
 import StartupScreen from './components/StartupScreen';
@@ -42,7 +41,7 @@ function AppContent({ isLoading }) {
         <AnimatePresence mode="wait">
           <Routes location={location} key={location.pathname}>
             <Route path="/" element={<FadeTransition><Home /></FadeTransition>} />
-            <Route path="/library" element={<FadeTransition><LibraryTesting /></FadeTransition>} />
+            <Route path="/library" element={<FadeTransition><Library /></FadeTransition>} />
             <Route path="/settings" element={<FadeTransition><Settings /></FadeTransition>} />
             <Route path="/harmony" element={<FadeTransition><Harmony /></FadeTransition>} />
           </Routes>
