--- jukeos/src/components/LibraryTesting.js
@@ -1,159 +1,0 @@
-import React, { useState, useRef } from 'react';
-import backgroundPng from '../assets/background.png';
-import mainGradient from '../assets/main-gradient.svg';
-import defaultAlbumArt from '../assets/default-album-art.png';
-import '../App.css';
-
-const ScrollWheel = ({ items }) => {
-  const [isDragging, setIsDragging] = useState(false);
-  const [startX, setStartX] = useState(0);
-  const [scrollLeft, setScrollLeft] = useState(0);
-  const wheelRef = useRef(null);
-
-  const handleMouseDown = (e) => {
-    setIsDragging(true);
-    setStartX(e.pageX - wheelRef.current.offsetLeft);
-    setScrollLeft(wheelRef.current.scrollLeft);
-  };
-
-  const handleMouseUp = () => {
-    setIsDragging(false);
-  };
-
-  const handleMouseMove = (e) => {
-    if (!isDragging) return;
-    e.preventDefault();
-    const x = e.pageX - wheelRef.current.offsetLeft;
-    const walk = (x - startX) * 2;
-    wheelRef.current.scrollLeft = scrollLeft - walk;
-  };
-
-  return (
-    <div 
-      className="scroll-wheel-container"
-      ref={wheelRef}
-      onMouseDown={handleMouseDown}
-      onMouseUp={handleMouseUp}
-      onMouseLeave={handleMouseUp}
-      onMouseMove={handleMouseMove}
-    >
-      <div className="scroll-wheel-track">
-        {items.map((item, index) => (
-          <div key={index} className="scroll-wheel-item">
-            <img 
-              src={item.imageUrl || defaultAlbumArt}
-              alt={item.title}
-              style={{
-                width: '150px',
-                height: '150px',
-                objectFit: 'cover',
-                borderRadius: '8px',
-                boxShadow: '0 4px 12px rgba(0,0,0,0.5)'
-              }}
-            />
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const LibrarySection = ({ title, items }) => {
-  return (
-    <div className="library-section">
-      <div className="section-title glow">
-        <h2>{title}</h2>
-      </div>
-      <div className="carousel-container">
-        <ScrollWheel items={items} />
-      </div>
-    </div>
-  );
-};
-
-const LibraryTesting = () => {
-  // TODO: Backend - Replace this mock data with actual Spotify API integration
-  // Required Spotify API endpoints:
-  // - GET /v1/me/playlists (for playlists)
-  // - GET /v1/me/shows (for podcasts)
-  // - GET /v1/me/albums (for albums)
-  // Consider implementing pagination if needed
-  const [sections] = useState([
-    {
-      title: 'PLAYLISTS',
-      items: Array(10).fill({ imageUrl: defaultAlbumArt, title: 'Playlist' })
-    },
-    {
-      title: 'PODCASTS',
-      items: Array(10).fill({ imageUrl: defaultAlbumArt, title: 'Podcast' })
-    },
-    {
-      title: 'ALBUMS',
-      items: Array(10).fill({ imageUrl: defaultAlbumArt, title: 'Album' })
-    }
-  ]);
-
-  // TODO: Backend - Add these states and effects for API integration:
-  // const [loading, setLoading] = useState(true);
-  // const { accessToken } = useContext(SpotifyAuthContext);
-  // useEffect(() => { fetch Spotify data here }, [accessToken]);
-
-  return (
-    <div style={{ position: 'relative' }}>
-      {/* TODO: Backend - Consider adding loading spinner/state here */}
-      <div style={{
-        backgroundImage: `url(${backgroundPng})`,
-        backgroundSize: 'cover',
-        backgroundPosition: 'center',
-        backgroundRepeat: 'no-repeat',
-        backgroundAttachment: 'fixed',
-        minHeight: '100vh',
-        padding: '20px',
-        perspective: '1000px',
-        overflow: 'hidden',
-        position: 'relative',
-        zIndex: 1
-      }}>
-        <div style={{
-          position: 'fixed',
-          bottom: 0,
-          left: '50%',
-          transform: 'translateX(-50%)',
-          width: '200vw',
-          height: 'auto',
-          zIndex: 2,
-          pointerEvents: 'none'
-        }}>
-          <img 
-            src={mainGradient} 
-            alt="gradient" 
-            style={{
-              width: '62%',
-              height: 'auto',
-              display: 'block',
-              margin: '0 auto'
-            }}
-          />
-        </div>
-        <div style={{
-          paddingTop: '120px',
-          paddingBottom: '40px',
-          overflowY: 'auto',
-          height: '100vh',
-          position: 'relative',
-          zIndex: 3
-        }}>
-          {sections.map((section, index) => (
-            <LibrarySection 
-              key={section.title}
-              title={section.title} 
-              items={section.items}
-            />
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default LibraryTesting;
